feat: :sparkles: Add Bookings Module connection
This module is useful to know if a flight has bookings

+++ src/main/java/co/udea/airline/api/services/flights/service/FlightService.java
@@ -3,6 +3,8 @@ package co.udea.airline.api.services.flights.service;
 import co.udea.airline.api.model.jpa.model.flights.Flight;
 import co.udea.airline.api.model.jpa.model.flights.Scale;
 import co.udea.airline.api.model.jpa.repository.flights.IFlightRepository;
+import co.udea.airline.api.services.bookings.service.BookingService;
+
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
@@ -17,6 +19,9 @@ public class FlightService {
 
     @Autowired
     private ScaleService scaleService;
+
+    @Autowired
+    private BookingService bookingService;
 
     public Flight saveFlight(Flight flight) {
 
@@ -58,10 +63,13 @@ public class FlightService {
     public Flight deleteFlightById(Long id) {
         Flight deletedFlight = getFlightById(id);
 
-        // TODO: Add booking service to validate if the flight has bookings
-
         if (deletedFlight == null) {
             return null;
+        }
+
+        boolean hasBookings = bookingService.flightHasBookings(id);
+        if (hasBookings) {
+            throw new IllegalArgumentException("The flight has bookings, you can't delete it.");
         }
 
         flightRepository.deleteById(id);
@@ -75,7 +83,12 @@ public class FlightService {
             return null;
         }
 
-        deleteFlightById(deletedFlight.getId());
+        boolean hasBookings = bookingService.flightHasBookings(deletedFlight.getId());
+        if (hasBookings) {
+            throw new IllegalArgumentException("The flight has bookings, you can't delete it.");
+        }
+
+        flightRepository.deleteById(deletedFlight.getId());
 
         return deletedFlight;
     }
